WHOOPS - This has been wrong since changeset 217:9cb0a4cb9afd!!
Basically, the *wrong* MACT headers, from the MACT ZIP, have been
commited so far. Instead, we shall use the headers from the sources
as released by 3DR with the Duke3D sources, as originally done
in a local working dir.

This (re)fixes the EDuke 2.00.23 build. In fact, not only
we get again the original EDuke EXE up to 3 differing bytes,
but the same also applies to the WW2GI EXE!