Let's move to the range-based APPVER_DN3DREV macro and related, as well as APPVER_EXEDEF, similarly to what was done for other games.
Note that the NAM, WW2 and EDUKE macros are *still* passed to the
compiler as arguments (along with APPVER_EXEDEF), and there a few places
in the code (especially the beginning of GAMEDEF.C) where this is vital.
Otherwise new definitions would have to be added to the very top,
possibly via an inclusion of the newly-added GAMEVER.H header.