FINALLY: Up to expected diffs, it's Duke Nukem 3D: Atomic Edition v1.5!
This has *really* been more time-confusing and nagging than it
should've. To summarize, except for a minor "APPVER" typo in
PLAYER.C:processinput (oh boy, this function again) and similarly
a different default weapons order in GAME.C:checkcommandline,
v1.5 has actually been FULLY ready in terms of code. Essentially,
we simply had to make sure that the NAM macro was *not* defined.

The gotcha, eventually, was figuring out which version of Watcom C
to use. You see, with v10.6, there was this minor difference in the
compiled GAME.C:domovethings function, and more.
Additionally, while v10.0b (and almost surely also 10.0a) was
somewhat closer in this function, it was worse elsewhere.
Eventually, after trying multiple combinations of optimizations for
wcc386 (using a bit of coding/scripting for automation of this),
Watcom C v10.5 was used, only to lead to the same issue as with 10.6.

Finally, it was decided to try something similar, but "in reverse".
Writing what has been known so far:
- As mentioned in the ROTT source code readme, 10.0b was used for ROTT
and the Apogee Sound System (ASS) v1.04. There's a very great evidence
for the latter in terms of the modification date of Watcom headers,
as present in AUDIO_WF.LIB from the released Duke3D sources,
which is v1.1.
- Furthermore, as just stated above, 10.0b was used for ASS v1.1,
a part of Duke3D v1.5 and EDuke 2.00.23.
- The Build engine object files, as well as MACT386.LIB objects
(and BUTWCD4.LIB objects) were apparently built with some form
of 10.5, again guessing from a modification date of Watcom headers.

What about the Duke3D v1.5 game code? Well, eventually, it looks like
README.TXT from 3DR's released sources was right again: Watcom C 10.0
was seemingly used. Not 10.6, 10.5, 10.0b or even 10.0a, but 10.0!

So, here we go. It's the same original binary DUKE3D.LE data, up to
a few gaps between C strings and co. The latter is apparently related
to the Watcom 10.0x series, since Matt's EXEs, made with 10.6, were
perfectly recreated (using the original binary OBJ/LIB files).