Move the definition of "count" so it comes right before ASTUB.C:Keys3d.
Not only this makes its location in the EXE relative to other global
variables closer to the original, it probably also makes since, since
the global "count" variable is essentially a replacement for local
variables of the same name.

Additionally, "timeoff" has unintentionally been relocated
in a similar manner, although it's a bit less stable
(evidenced by bits of experimenting). There's also
a gap at the wrong location. This gaps seems
to come from the unused copy of "hvel".