Few more hacks, especially the "unusedvar" variable. On the one hand, it is defined on the top, so there's the addition of a 2-bytes gap between temppicnum and lockbyte4094 in the EXE.
On the other hand, to ensure that timeoff is still in-between yvel
and xvel, searchit was omitted (it's been kinda tricky with timeoff).

Ignoring the remaining known few differences in Keys3d_,
and other few differences, the recreated EXE's code and
data are almost identical to the original's at this point.