duke3d/ASTUB.C: Looks like the code involving the calls to initengine is the same in v1.3d's editor as it is in v1.4.
It's possible the CTW modification from the GPLed release of ASTUB.C
covers a few more changes, which aren't directly shown in the file, and
were reverted for the recreation of v1.4 (e.g., chainxres, chainyres).

While BUILD2.TXT mentions the removal of support for specially optimized
modes on 9/25/96, v1.4's Build Editor still has loopnumofsector
and setfirstwall in BUILD.OBJ, mentioned to be moved
into ENGINE.OBJ only in the 9/6/96 entry.

At least in terms of function sizes, after reverting these few changes,
the GPLed ASTUB.C seems to match v1.3d. Most of the recreation work
is expected to be done in the Build Engine codebase, either way.